Refactor login flow and update Prisma adapter
Moved the login form from login.vue to index.vue and removed login.vue. Added homepage.vue to display user info after login. Updated prisma.ts to use @prisma/adapter-better-sqlite3 and added the adapter dependency in package.json for improved SQLite support.
