+++ WWM-IMPACT-CRM/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@nuxtjs/tailwindcss": "^6.14.0",
+    "@prisma/adapter-better-sqlite3": "^6.18.0",
     "@prisma/client": "^6.16.3",
     "@prisma/nuxt": "^0.3.0",
     "@vue/runtime-core": "^3.5.22",
